pyproject.toml
[project]
name = "transactions"
version = "0.1.0"
description = "overviewing transaction isolation in postgres"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
]
